Added basic pvalve tool.
This is a "pipe valve" where rate of data flowing from stdin to stdout
is limited using the existing leaky bucket crate.

While running, you edit the file "control" in the current directory and
put a number in the first line and it will limit the amount of bytes to
pass through the filter per second to that value.

The only thing this offers over pv's rate limiting feature is that it
can change the maximum speed during runtime. It is currently not able to
operate in "line mode" which is a good feature to add when working with
line-oriented structured data rather than raw bytes.