syncio: Added optimizations.
When the cost of the current buffer fits within what was granted by the
rate limiter, the original buffer can be transferred.

One more possible optimization that can be made is to actually size the
buffer as whatever fits before cost + 1 which might allow the whole
buffer to pass if the remainder of the buffer has a cost of 0.